rayrender 0.11.0: Microfacets + metal update
-`microfacet()` Add microfacet (anisotropic + isotropic) material
-`metal()` Add physically-based reflection and attenuation coefficients to metals, as well as procedural and other texturing options
-Fix implementation of Oren-Nayar diffuse material
-`generate_cornell()` Add sigma option to cornell box walls

-Fix bug with ray origins
-Fix surface acne when camera far away from intersection point (see "A Fast and Robust Method for Avoiding Self-Intersection", Carsten Wächter and Nikolaus Binder)

+++ src/mathinline.h
@@ -171,7 +171,7 @@ inline bool SameHemisphere(const vec3 &w, const vec3 &wp) {
 
 
 inline Float AbsDot(const vec3 &v1, const vec3 &v2) {
-  return(std::fabsf(dot(v1,v2)));
+  return(std::fabs(dot(v1,v2)));
 } 
 
 template <typename IN_T, typename OUT_T>
